+++ src/monet.d.ts
@@ -32,6 +32,11 @@ interface Bind<T> extends Chain<T> {
   join<V>(): Bind<V>  // works only if T = Bind<V>
 }
 
+/* Typeclass for catamorphism */
+interface Catamorphism<F, T> {
+  cata<C>(l: (e?: F) => C, r: (v: T) => C): C;
+}
+
 /* Typeclass for traversables */
 export interface ITraversable<T> {
   foldLeft<V>(initial: V): (fn: (acc: V, val: T) => V) => V;
@@ -53,6 +58,7 @@ interface IMonad<T> extends Functor<T>, Bind<T>, Applicative<T> {
   /* These are monet-Monad-specific: */
   takeLeft(m: IMonad<T>): IMonad<T>;
   takeRight(m: IMonad<T>): IMonad<T>;
+  ['@@type']: string;
 }
 
 interface IMonadFactory extends Function {
@@ -96,6 +102,8 @@ interface IIdentityStatic extends IIdentityFactory, IMonadStatic {
   unit: IIdentityFactory;
   of: IIdentityFactory;    // alias for unit
   pure: IIdentityFactory;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Identity<any>;
 }
 
 export const Identity: IIdentityStatic;
@@ -104,7 +112,8 @@ export const Identity: IIdentityStatic;
  * Maybe
  */
 
-export interface Maybe<T> extends IMonad<T>, Setoid<Maybe<T>>, ITraversable<T> {
+export interface Maybe<T>
+  extends IMonad<T>, Setoid<Maybe<T>>, ITraversable<T>, Catamorphism<undefined, T> {
   /* Inherited from Monad: */
   bind<V>(fn: (val: T) => Maybe<V>): Maybe<V>;
   flatMap<V>(fn: (val: T) => Maybe<V>): Maybe<V>;
@@ -161,9 +170,12 @@ interface IMaybeStatic extends IMonadStatic {
   Nothing: INoneStatic;
   fromFalsy<V>(val: V|null|undefined): Maybe<V>;
   fromNull<V>(val: V|null|undefined): Maybe<V>;
+  fromUndefined<V>(val: V|undefined): Maybe<V>;
   unit: ISomeStatic;
   of: ISomeStatic;    // alias for unit
   pure: ISomeStatic;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Maybe<any>;
 }
 
 export const Some: ISomeStatic;
@@ -176,7 +188,8 @@ export const Maybe: IMaybeStatic;
  * Either
  */
 
-export interface Either<E, T> extends IMonad<T>, Setoid<Either<E, T>>, ITraversable<T> {
+export interface Either<E, T>
+  extends IMonad<T>, Setoid<Either<E, T>>, ITraversable<T>, Catamorphism<E, T> {
   /* Inherited from Monad: */
   bind<V>(fn: (val: T) => Either<E, V>): Either<E, V>;
   flatMap<V>(fn: (val: T) => Either<E, V>): Either<E, V>;
@@ -216,6 +229,8 @@ interface IEitherStatic extends IMonadStatic {
   unit: IRightStatic;
   of: IRightStatic;    // alias for unit
   pure: IRightStatic;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Either<any, any>;
 }
 
 interface IRightStatic extends IMonadFactory {
@@ -238,7 +253,8 @@ interface IValidationAcc extends Function {
   (): IValidationAcc;
 }
 
-export interface Validation<E, T> extends IMonad<T>, Setoid<Validation<E, T>>, ITraversable<T> {
+export interface Validation<E, T>
+  extends IMonad<T>, Setoid<Validation<E, T>>, ITraversable<T>, Catamorphism<E, T> {
   /* Inherited from Monad: */
   bind<V>(fn: (val: T) => Validation<E, V>): Validation<E, V>;
   flatMap<V>(fn: (val: T) => Validation<E, V>): Validation<E, V>;
@@ -281,6 +297,8 @@ interface IValidationStatic extends IMonadStatic {
   of: ISuccessStatic;     // alias for unit
   pure: ISuccessStatic;   // alias for unit
   point: ISuccessStatic;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Validation<any, any>;
 }
 
 interface ISuccessStatic extends IMonadFactory {
@@ -362,6 +380,8 @@ interface IListStatic extends IMonadStatic {
   unit: IListFactory;
   of: IListFactory;    // alias for unit
   pure: IListFactory;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is List<any>;
 }
 
 export const List: IListStatic;
@@ -428,6 +448,8 @@ interface INELStatic extends INELFactory, IMonadStatic {
   unit: INELFactory;
   of: INELFactory;    // alias for unit
   pure: INELFactory;  // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is NEL<any>;
 }
 
 export const NonEmptyList: INELStatic;
@@ -465,6 +487,8 @@ interface IIOStatic extends IIOFactory, IMonadStatic {
   of: IIOFactory;    // alias for unit
   pure: IIOFactory;  // alias for unit
   io: IIOFactory;    // alias for unit
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is IO<any>;
 }
 
 export const IO: IIOStatic;
@@ -499,6 +523,8 @@ interface IReaderStatic extends IReaderFactory, IMonadStatic {
   pure: IReaderFactory;  // alias for unit
   point: IReaderFactory; // alias for unit
   ask<E>(): Reader<E, E>;
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Reader<any, any>;
 }
 
 export const Reader: IReaderStatic;
@@ -545,6 +571,8 @@ interface IFreeStatic extends IMonadStatic {
   of: IReturnStatic;    // alias for unit
   pure: IReturnStatic;  // alias for unit
   liftF<A, FA>(fa: FA): Free<A>; // FA = F<A>
+  isOfType(target: any): boolean;
+  isInstance(target: any): target is Free<any>;
 }
 
 interface IReturnStatic extends IMonadFactory {